Add default command tests for action based subcommand
